Unify the color radio bootloader UI into a single source file and move to the gui folder.
